+++ content_scripts/tranquility_event_handlers.js
@@ -4,7 +4,7 @@
  * cluttered web pages
  **********************************************************************
 
-   Copyright (c) 2012-2019 Arun Kunchithapatham
+   Copyright (c) 2012-2020 Arun Kunchithapatham
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -224,6 +224,30 @@ function handleImportInputClickEvent(event) {
     reader.readAsText(input_file);
 }
 
+function handleLoadOriginalLinkClickEvent(event) {
+    event.preventDefault();
+    event.stopPropagation();
+    window.location.reload();
+}
+
+function handleSaveAsPDFClickEvent(event) {
+    console.log("Got into Save As PDF click event handler");
+    event.preventDefault();
+    event.stopPropagation();
+    console.log("Stopped Propagation");
+    window.print();
+}
+
+function handleShowPreferencesClickEvent(event) {
+    console.log("Got into Show Preferences click event handler");
+    event.preventDefault();
+    event.stopPropagation();
+    console.log("Stopped Propagation");
+    browser.runtime.sendMessage(
+    {
+     "action": "openOptionsPage"
+    });
+}
 
 function handleClickEvent(event) {
 
@@ -237,7 +261,17 @@ function handleClickEvent(event) {
         // and continue
         //
         if (urlStr.split("#")[0] == currentURL.split("#")[0]) {
-            console.log("Do nothing - we want to navigate to an anchor in the current page...");
+            // If we have clicked on the original_link anchor
+            // and really want to return to the original webpage,
+            // then load the page again (after stripping the #)
+            if (event.target.className == 'tranquility_original_link_anchor') {
+                event.preventDefault();
+                event.stopPropagation();
+                window.location.assign(urlStr.split("#")[0]);
+            }
+            else {
+                console.log("Do nothing - we want to navigate to an anchor in the current page...");
+            }
         }
         // Else...
         // Do not load link - instead request background page to load it and then
@@ -325,7 +359,9 @@ function addBackEventListeners() {
         "tranquility_view_notes"            : handleAnnotationViewOrNoteClickEvent,
         "tranquility_delete_offline_link"   : handleDeleteOfflineLinkClickEvent,
         "tranquility_offline_link"          : handleLoadOfflineLinkClickEvent,
-        "tranquility_expand_menu_btn"       : handleExpandMenuButtonClickEvent
+        "tranquility_expand_menu_btn"       : handleExpandMenuButtonClickEvent,
+        "tranquility_saveaspdf_div"         : handleSaveAsPDFClickEvent,
+        "tranquility_prefs_link_div"        : handleShowPreferencesClickEvent
 
     };